requirements.txt
pycairo==1.27.0
pygame==2.6.1
